tools/requirements.txt
lxml
markdownify
requests
pyyaml
pillow
